+++ coursework/include/game/entity/component/gun/LaunchData.hpp
@@ -6,8 +6,8 @@ struct LaunchData {
     sf::Vector2f point;
     sf::Vector2f velocity;
     float angle;
-    int num;
+    size_t num;
     LaunchData(const BulletMultipliers &bullet_multipliers, const sf::Vector2f &point,
-            const sf::Vector2f &velocity, const float angle, const int num):
+            const sf::Vector2f &velocity, const float angle, const size_t num):
         bullet_multipliers(bullet_multipliers), point(point), velocity(velocity), angle(angle), num(num) {}
 };
