+++ coursework/include/sprite/Infos.hpp
@@ -29,7 +29,7 @@ class SpriteInfos final {
     using InitializerInfo = InitializerInfo<AbstractInfoVector>;
     using InitList = InitializerList<InitializerInfo>;
 
-    [[nodiscard]] const InfoVector& addInfoFromAbstract(const size_t index, const int num);
+    [[nodiscard]] const InfoVector& addInfoFromAbstract(const size_t index, const size_t num);
     void handleTextureLoad(const size_t index, const sf::Vector2i &texture_load_range);
 
     AbstractInfos _abstract_infos;
@@ -37,8 +37,8 @@ class SpriteInfos final {
 public:
     SpriteInfos(const InitList &init_list);
 
-    [[nodiscard]] const InfoVector& getInfo(const ElementType element_type, const int num);
-    void addInfoFromAbstract(const ElementType element_type, const int num);
+    [[nodiscard]] const InfoVector& getInfo(const ElementType element_type, const size_t num);
+    void addInfoFromAbstract(const ElementType element_type, const size_t num);
 
     ~SpriteInfos() noexcept = default;
     SpriteInfos(SpriteInfos&&) noexcept = default;
